requirements.txt
eel==0.14.0
python-i18n[YAML]==0.3.9
screeninfo==0.8.1
